SVN_MERGE
Merged revisions 1039402,1039529,1039541,1039550,1039613,1039615,1039618,1039720,1039726,1039729-1039730,1039737,1039739,1039775,1039937,1039940,1040037,1040091 via svnmerge from 
https://vkrause@svn.kde.org/home/kde/trunk/KDE/kdepimlibs

................
  r1039402 | skelly | 2009-10-23 12:12:15 +0200 (Fri, 23 Oct 2009) | 1 line
  
  Use the isReadOnly method
................
  r1039529 | tnyblom | 2009-10-23 20:11:23 +0200 (Fri, 23 Oct 2009) | 5 lines
  
  Change KIO::ERR_ values to better fit the messages given.
  
  CCBUG: 166621
  CCBUG: 211050
................
  r1039541 | winterz | 2009-10-23 21:28:22 +0200 (Fri, 23 Oct 2009) | 9 lines
  
  when declining, don't complain about a missing invitation in your calendar if you
  haven't put it there yet.
  kolab/issue3922
  
  also, use information dialogs instead of error dialogs to tell the user that
  the invitation is missing from their calendar.
  
  MERGE: e35,e4,4.3
................
  r1039550 | winterz | 2009-10-23 21:50:22 +0200 (Fri, 23 Oct 2009) | 22 lines
  
  Blocked revisions 1039547 via svnmerge
  
  ................
    r1039547 | winterz | 2009-10-23 15:46:53 -0400 (Fri, 23 Oct 2009) | 16 lines
    
    Merged revisions 1039545 via svnmerge from 
    https://svn.kde.org/home/kde/branches/kdepim/enterprise/kdepim
    
    ........
      r1039545 | winterz | 2009-10-23 15:37:39 -0400 (Fri, 23 Oct 2009) | 9 lines
      
      when declining, don't complain about a missing invitation in your calendar if you
      haven't put it there yet.
      kolab/issue3922
      
      also, use information dialogs instead of error dialogs to tell the user that
      the invitation is missing from their calendar.
      
      MERGE: e4 (already in trunk)
    ........
  ................
................
  r1039613 | winterz | 2009-10-24 01:24:35 +0200 (Sat, 24 Oct 2009) | 3 lines
  
  compile warnings--
  MERGE: none
................
  r1039615 | winterz | 2009-10-24 01:40:32 +0200 (Sat, 24 Oct 2009) | 4 lines
  
  compile warning--
  remove a few unneeded endls in kDebugs()
  MERGE: none
................
  r1039618 | winterz | 2009-10-24 01:45:36 +0200 (Sat, 24 Oct 2009) | 3 lines
  
  revertlast.
  mistakenly had a reviewboard patch in here.
................
  r1039720 | tnyblom | 2009-10-24 14:30:30 +0200 (Sat, 24 Oct 2009) | 4 lines
  
  Improve error messages.
  
  CCBUG: 166621
................
  r1039726 | edulix | 2009-10-24 14:48:46 +0200 (Sat, 24 Oct 2009) | 5 lines
  
  Fixing some incorrectly written endifs in GpgmeppConfig.cmake.in which gave some warnings and a typo (Gpgmepp vs. gpgmepp/) in an include() directive which made find_package(Gpgmepp REQUIRED) fail. I suppose I should backport this too?
  
  CCMAIL:kde-pim@kde.org
................
  r1039729 | winterz | 2009-10-24 15:19:18 +0200 (Sat, 24 Oct 2009) | 2 lines
  
  some easy Krazy fixes: non-const dpointer; explicit ctor; spelling; includes
................
  r1039730 | winterz | 2009-10-24 15:21:47 +0200 (Sat, 24 Oct 2009) | 3 lines
  
  fix passing an arg to i18n.
  give i18n context on a couple strings for the translators
................
  r1039737 | winterz | 2009-10-24 15:36:15 +0200 (Sat, 24 Oct 2009) | 3 lines
  
  use KLineEdit instead of QLineEdit
  use KComboBox instead of QComboBox
................
  r1039739 | winterz | 2009-10-24 15:42:56 +0200 (Sat, 24 Oct 2009) | 3 lines
  
  minor coding style
  SVN_SILENT:
................
  r1039775 | coles | 2009-10-24 17:26:45 +0200 (Sat, 24 Oct 2009) | 3 lines
  
  Proof-reading - removed strings that although jocular, we shouldn't risk being seen by users.
................
  r1039937 | winterz | 2009-10-25 01:14:25 +0200 (Sun, 25 Oct 2009) | 3 lines
  
  compile warnings--
  MERGE: none
................
  r1039940 | winterz | 2009-10-25 01:39:08 +0200 (Sun, 25 Oct 2009) | 3 lines
  
  add new typeStr() methods for that return translated strings of the incidence types.
  MERGE: none
................
  r1040037 | tokoe | 2009-10-25 12:35:54 +0100 (Sun, 25 Oct 2009) | 5 lines
  
  Close the editor in the Name column when user selected
  name from autocompletion popup menu
  
  BUG: 211577
................
  r1040091 | winterz | 2009-10-25 16:09:35 +0100 (Sun, 25 Oct 2009) | 3 lines
  
  revert BIC addition of typeStr().
  comment-out typeStr() for implementing in KDE5
................

svn path=/branches/work/akonadi-ports/kdepimlibs/; revision=1040444

+++ kcal/alarm.cpp
@@ -486,7 +486,7 @@ KDateTime Alarm::nextRepetition( const KDateTime &preTime ) const
     // there isn't an occurrence after the specified time
     return KDateTime();
   }
-  int repetition;
+  qint64 repetition;
   int interval = d->mAlarmSnoozeTime.value();
   bool daily = d->mAlarmSnoozeTime.isDaily();
   if ( daily ) {
@@ -502,7 +502,7 @@ KDateTime Alarm::nextRepetition( const KDateTime &preTime ) const
     // all repetitions have finished before the specified time
     return KDateTime();
   }
-  return daily ? at.addDays( repetition * interval )
+  return daily ? at.addDays( int( repetition * interval ) )
                : at.addSecs( repetition * interval );
 }
 
@@ -516,7 +516,7 @@ KDateTime Alarm::previousRepetition( const KDateTime &afterTime ) const
   if ( !d->mAlarmRepeatCount ) {
     return at;
   }
-  int repetition;
+  qint64 repetition;
   int interval = d->mAlarmSnoozeTime.value();
   bool daily = d->mAlarmSnoozeTime.isDaily();
   if ( daily ) {
@@ -531,7 +531,7 @@ KDateTime Alarm::previousRepetition( const KDateTime &afterTime ) const
   if ( repetition > d->mAlarmRepeatCount ) {
     repetition = d->mAlarmRepeatCount;
   }
-  return daily ? at.addDays( repetition * interval )
+  return daily ? at.addDays( int( repetition * interval ) )
                : at.addSecs( repetition * interval );
 }
